:recycle: Refactor logic for redis queue naming on each limited method
We were using by default a queue name "queue" for the redis-based queues. Now we will defaultly name it using the Class and the mehod name to try to make it more unique in cases where the user doesn't really need to bother in passing a key.

This way, the redis key is different for each class+method limited. For more specific cases, the user can still pass their own queue value as keyword argument.

 Also, we now support to send a `key` keyword argument so the limited method queue name can be customized.